+++ .idea/Gateway.iml
@@ -6,5 +6,6 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="Practice_IotGateway" />
   </component>
 </module>
